+++ firmware_arduino/lib/PengolahSinyalPPG/src/cariLembah.h
@@ -4,11 +4,11 @@
 #include <Arduino.h>
 
 // Deklarasi fungsi (hanya judulnya saja)
-void cariLembah(uint16_t *sinyalBuffer, uint8_t panjangBuffer,
+void cariLembah(float *sinyalBuffer, uint8_t panjangBuffer,
                 uint8_t minProminence, uint8_t minDistance,
-                uint16_t *outPeakValues, uint8_t *outPeakLocs,
+                float *outPeakValues, uint8_t *outPeakLocs,
                 uint8_t *outNumPeaks);
-uint8_t kandidatLembah(uint16_t *sinyalBuffer, uint8_t panjangBuffer,
+uint8_t kandidatLembah(float *sinyalBuffer, uint8_t panjangBuffer,
                        uint8_t *indeks);
 
 #endif
