+++ firmware_arduino/lib/PengolahSinyalPPG/src/cariPuncak.cpp
@@ -1,14 +1,14 @@
 #include "cariPuncak.h"
 
-void cariPuncak(uint16_t *sinyalBuffer, uint8_t panjangBuffer,
+void cariPuncak(float *sinyalBuffer, uint8_t panjangBuffer,
                 uint8_t minProminence, uint8_t minDistance,
-                uint16_t *outPeakValues, uint8_t *outPeakLocs,
+                float *outPeakValues, uint8_t *outPeakLocs,
                 uint8_t *outNumPeaks) {
   // --- LANGKAH 0: Hitung Rata-rata (Baseline) untuk Zero Crossing ---
-  uint32_t totalSum = 0;
+  float totalSum = 0;
   for (uint8_t i = 0; i < panjangBuffer; i++)
     totalSum += sinyalBuffer[i];
-  uint32_t mean = totalSum / (float)panjangBuffer;
+  float mean = totalSum / (float)panjangBuffer;
 
   uint8_t zcCount = 0;
   for (uint8_t i = 1; i < panjangBuffer; i++) {
@@ -95,7 +95,7 @@ void cariPuncak(uint16_t *sinyalBuffer, uint8_t panjangBuffer,
   *outNumPeaks = validCount; // Kembalikan jumlah total puncak yang ditemukan
 }
 
-uint8_t kandidatPuncak(uint16_t *sinyalBuffer, uint8_t panjangBuffer,
+uint8_t kandidatPuncak(float *sinyalBuffer, uint8_t panjangBuffer,
                        uint8_t *indeks) {
   uint8_t count = 0;
   uint8_t gap = 1; // Jarak perbandingan kiri-kanan
