+++ firmware_arduino/lib/PengolahSinyalPPG/src/cariPuncak.h
@@ -3,11 +3,11 @@
 
 #include <Arduino.h>
 
-void cariPuncak(uint16_t *sinyalBuffer, uint8_t panjangBuffer,
+void cariPuncak(float *sinyalBuffer, uint8_t panjangBuffer,
                 uint8_t minProminence, uint8_t minDistance,
-                uint16_t *outPeakValues, uint8_t *outPeakLocs,
+                float *outPeakValues, uint8_t *outPeakLocs,
                 uint8_t *outNumPeaks);
-uint8_t kandidatPuncak(uint16_t *sinyalBuffer, uint8_t panjangBuffer,
+uint8_t kandidatPuncak(float *sinyalBuffer, uint8_t panjangBuffer,
                        uint8_t *indeks);
 
 #endif
